Move SiteCorrections into pipeline_utils and setup v2
